pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strands-live"
version = "0.1.0"
description = "Live speech agent using Amazon Nova Sonic with Strands SDK"
authors = [{name = "murmeral"}]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[bedrock]>=0.1.0",
    "strands-agents-tools>=0.1.0", 
    "pyaudio>=0.2.11",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
strands-live = "strands_live.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"